refactor: extension UITextView 위치 수정 및 접근 제어자 수정
